requirements.txt
Django
django-tastypie
git+ssh://git@github.com/getsentry/raven-python.git
